apply op logic works
