+++ src/main/scala/somind/dtlab/routes/ActorApiRoute.scala
@@ -1,13 +1,9 @@
 package somind.dtlab.routes
 
-import akka.http.scaladsl.model._
 import akka.http.scaladsl.server._
-import akka.pattern._
 import com.typesafe.scalalogging.LazyLogging
-import somind.dtlab.Conf._
 import somind.dtlab.HttpSupport
 import somind.dtlab.models._
-import somind.dtlab.observe.Observer
 import somind.dtlab.routes.functions._
 
 import scala.language.postfixOps
@@ -39,22 +35,6 @@ object ActorApiRoute
     with StateApiTrait
     with GetChildNamesTrait
     with Directives {
-
-  def applyTelemetryMsg(dtp: DtPath, telemetry: Telemetry): Route = {
-    onSuccess(dtDirectory ask TelemetryMsg(dtp, telemetry)) {
-      case DtOk() =>
-        Observer("actor_route_post_success")
-        complete(StatusCodes.Accepted)
-      case DtErr(emsg) =>
-        Observer("actor_route_post_failure")
-        logger.debug(s"unable to post telemetry: $emsg")
-        complete(StatusCodes.BadRequest, emsg)
-      case e =>
-        Observer("actor_route_post_unk_err")
-        logger.warn(s"unable to handle: $e")
-        complete(StatusCodes.InternalServerError)
-    }
-  }
 
   private def applyProps(segs: List[String],
                          limit: Option[Int],
