+++ src/main/scala/somind/dtlab/routes/functions/TelemetryUnWrappers.scala
@@ -2,14 +2,34 @@ package somind.dtlab.routes.functions
 
 import akka.http.scaladsl.model._
 import akka.http.scaladsl.server._
+import akka.pattern.ask
 import com.typesafe.scalalogging.LazyLogging
+import somind.dtlab.Conf._
 import somind.dtlab.models._
-import somind.dtlab.routes.ActorApiRoute.applyTelemetryMsg
+import somind.dtlab.observe.Observer
 import somind.dtlab.routes.functions.TelemetryUnMarshallers._
 
-object TelemetryUnWrappers extends JsonSupport with Directives with LazyLogging {
+object TelemetryUnWrappers
+    extends JsonSupport
+    with Directives
+    with LazyLogging {
 
   type UnWrapper = DtPath => Route
+  def applyTelemetryMsg(dtp: DtPath, telemetry: Telemetry): Route = {
+    onSuccess(dtDirectory ask TelemetryMsg(dtp, telemetry)) {
+      case DtOk() =>
+        Observer("actor_route_post_success")
+        complete(StatusCodes.Accepted)
+      case DtErr(emsg) =>
+        Observer("actor_route_post_failure")
+        logger.debug(s"unable to post telemetry: $emsg")
+        complete(StatusCodes.BadRequest, emsg)
+      case e =>
+        Observer("actor_route_post_unk_err")
+        logger.warn(s"unable to handle: $e")
+        complete(StatusCodes.InternalServerError)
+    }
+  }
 
   def NamedUnWrapper(dtp: DtPath): Route = {
     entity(as[LazyNamedTelemetry]) { ntelem =>
